requirements.txt
numpy~=1.26.4
pandas~=2.2.3

openai~=1.62.0
streamlit~=1.42.0
python-dotenv~=1.0.1